LAB6_MidtermLab/BookManager/backend/requirements.txt
alembic==1.14.1
bcrypt==4.0.1
fastapi==0.115.12
passlib==1.7.4
pydantic==2.10.6
pydantic[email]==2.10.6
pydantic-settings==2.8.1
python-jose==3.4.0
pytz==2025.2
sqlalchemy==2.0.40
uvicorn==0.33.0